Fundamental restructuring of the code into API, CLI, spec and scripts
