+++ src/main/kotlin/me/vendoor/dragonball/specification/VendoorDatabaseSpecification.kt
@@ -1,7 +1,7 @@
 package me.vendoor.dragonball.specification
 
 import com.typesafe.config.Config
-import me.vendoor.dragonball.dsl.database
+import me.vendoor.dragonball.api.dsl.database
 
 fun vendoorDatabaseSpecification(config: Config) = database {
     name { config.getString("database.name") }
